pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lofi-generator"
version = "0.1.0"
description = "Production-ready AI-powered lo-fi music generation system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Lo-Fi Generator Team"}
]
keywords = ["music-generation", "ai", "lofi", "gpt2", "midi", "audio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.12.0",
    "accelerate>=0.20.0",
    "miditok>=2.1.0",
    "miditoolkit>=1.0.0",
    "pretty_midi>=0.2.10",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "pydub>=0.25.0",
    "pedalboard>=0.7.0",
    "tensorboard>=2.13.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyloudnorm>=0.1.1",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.3.0",
    "mypy>=1.3.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "flake8-docstrings>=1.7.0",
    "pylint>=2.17.0",
    "bandit>=1.7.0",
    "pre-commit>=3.3.0",
]
docs = [
    "sphinx>=6.2.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.23.0",
    "myst-parser>=2.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.23.0",
    "nbformat>=5.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/lofi-generator"
Documentation = "https://lofi-generator.readthedocs.io"
Repository = "https://github.com/yourusername/lofi-generator"
"Bug Tracker" = "https://github.com/yourusername/lofi-generator/issues"

[project.scripts]
lofi-tokenize = "scripts.01_tokenize:main"
lofi-build-dataset = "scripts.02_build_dataset:main"
lofi-train = "scripts.03_train:main"
lofi-generate = "scripts.04_generate:main"
lofi-batch-generate = "scripts.05_batch_generate:main"

[tool.setuptools]
packages = ["src", "scripts"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

# mypy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "miditok.*",
    "miditoolkit.*",
    "pretty_midi.*",
    "librosa.*",
    "pedalboard.*",
    "pyloudnorm.*",
]
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "requires_gpu: Tests requiring GPU",
    "requires_fluidsynth: Tests requiring FluidSynth",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
skip_empty = true
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Pylint configuration
[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
]

[tool.pylint.design]
max-args = 10
max-attributes = 15
max-bool-expr = 5
max-branches = 15
max-locals = 20
max-parents = 7
max-public-methods = 25
max-returns = 8
max-statements = 60
min-public-methods = 1

# Bandit security linter configuration
[tool.bandit]
exclude_dirs = ["tests", "venv"]
skips = ["B101"]  # assert_used
